pyproject.toml
[project]
name = "engineering-team"
version = "0.1.0"
description = "Multi-agent engineering team built with Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "pyyaml>=6.0.0",
    "gradio>=4.0.0",
    "pytest>=7.0.0",
]
